refactor(dashboard): extract Court policies + bindings sub-routers (F-B-202 PR-8/10)
Audit ref: imp/audits/2026-05-20/findings/track-b/F-B-202.md

Sprint 2 modularization PR-8 of 10. base=main.

Extracts 8 routes (4 policies + 4 bindings) from `app/dashboard/router.py`
into two new sibling modules. Pure cut+paste; behaviour-equivalent.

  app/dashboard/policies_routes.py (new, 168 LOC, 4 routes):
    GET  /dashboard/policies                          list session policies
    GET  /dashboard/policies/create                   create form
    POST /dashboard/policies/create                   submit new policy
    POST /dashboard/policies/{id}/deactivate          flip is_active=False

  app/dashboard/bindings_routes.py (new, 222 LOC, 4 routes):
    GET  /dashboard/bindings                          list (org × agent grants)
    POST /dashboard/bindings/{id}/approve             approve pending binding
    POST /dashboard/bindings/{id}/revoke              revoke + close sessions
    POST /dashboard/bindings/{id}/scope               edit scope (subset check)

  app/dashboard/router.py (2071 → 1778 LOC, -293):
    - Sections 25 (Policies) + 26 (Bindings) removed; breadcrumb comments
      at the old positions.
    - Two new include_router calls.
    - Unused imports trimmed: PolicyRecord + create_policy + get_policy +
      deactivate_policy (now in policies_routes.py); get_binding +
      list_all_bindings + update_binding_scope (now in bindings_routes.py).

Behavioural delta: zero. CSRF + per-org gate on policy_deactivate +
binding revoke side-effects (close sessions + drop ws) preserved.

Pattern follows orgs_routes.py / auth_routes.py: sub-router declares
``APIRouter(tags=[...])`` with **no prefix**, route paths bare
(``/policies`` / ``/bindings``); the outer ``/dashboard`` prefix is
prepended by ``router.include_router(...)`` in router.py — matches the
PR-2 footnote convention.

Tests:
  - pytest tests/test_dashboard_bindings.py — 11/11
  - pytest tests/test_dashboard*.py tests/test_policy*.py
    tests/test_onboarding.py tests/test_enrollment.py — 252/252
    (1 pre-existing env-fail deselected: test_badge_approvals_*
    requires cullis_enterprise NOT installed, which it is locally
    for dual-repo work; passes in CI).

Router LOC progress:
  Pre-Sprint 2 : 67 routes, 3125 LOC
  Post-PR-1 #839: 67 routes, 2930 LOC   merged
  Post-PR-2 #841: 60 routes, 2614 LOC   merged
  Post-PR-3 #842: 57 routes, 2528 LOC   merged
  Post-PR-4 #847: 50 routes, 2372 LOC   merged
  Post-PR-5 #848: 39 routes, 2071 LOC   merged
  Post-PR-8 (this): 31 routes, 1778 LOC
  Target post-PR-10: ~3 (aggregator), ~100 LOC

Conflict warning: PR-6 #849 (org_onboard+seal) and PR-7 #850 (agents)
are still open. Sections in this PR (Policies 25 + Bindings 26) have
zero line overlap with their sections — rebase clean post-merge in any
order.

Next PR-9 (F-B-202 modularization): sessions + audit + rfq sub-routers
(sections 13+14+15, ~6 routes, ~275 LOC).
