+++ app/dashboard/router.py
@@ -59,9 +59,8 @@ from app.registry.org_store import (
 )
 from app.registry.binding_store import (
     BindingRecord, create_binding, approve_binding, revoke_binding,
-    get_binding_by_org_agent, get_binding, list_all_bindings, update_binding_scope,
+    get_binding_by_org_agent,
 )
-from app.policy.store import PolicyRecord, create_policy, get_policy, deactivate_policy
 from app.broker.db_models import SessionRecord, SessionMessageRecord, RfqRecord, RfqResponseRecord
 from app.broker.ws_manager import ws_manager
 from app.auth.transaction_token import create_transaction_token, compute_payload_hash
@@ -119,6 +118,13 @@ from app.dashboard import agents_lifecycle_routes as _agents_lifecycle_routes  #
 from app.dashboard import agents_credentials_routes as _agents_credentials_routes  # noqa: E402
 router.include_router(_agents_lifecycle_routes.router)
 router.include_router(_agents_credentials_routes.router)
+
+# F-B-202 PR-8: include policies (session policy CRUD) and bindings
+# (cross-org grant registry). 8 routes total.
+from app.dashboard import policies_routes as _policies_routes  # noqa: E402
+from app.dashboard import bindings_routes as _bindings_routes  # noqa: E402
+router.include_router(_policies_routes.router)
+router.include_router(_bindings_routes.router)
 
 
 # ─────────────────────────────────────────────────────────────────────────────
@@ -622,308 +628,8 @@ async def rfq_approve(request: Request, rfq_id: str, db: AsyncSession = Depends(
 
 
 
-# ─────────────────────────────────────────────────────────────────────────────
-# Policies
-# ─────────────────────────────────────────────────────────────────────────────
-
-@router.get("/policies", response_class=HTMLResponse)
-async def policies_list(request: Request, db: AsyncSession = Depends(get_db)):
-    session = require_login(request)
-    if isinstance(session, RedirectResponse):
-        return session
-
-    result = await db.execute(
-        select(PolicyRecord)
-        .where(PolicyRecord.policy_type == "session")
-        .order_by(PolicyRecord.org_id, PolicyRecord.policy_id)
-    )
-    records = result.scalars().all()
-
-    policy_list = []
-    for r in records:
-        conds = r.rules.get("conditions", {})
-        policy_list.append({
-            "policy_id": r.policy_id,
-            "org_id": r.org_id,
-            "target_orgs": conds.get("target_org_id", []),
-            "capabilities": conds.get("capabilities", []),
-            "effect": r.rules.get("effect", "allow"),
-            "is_active": r.is_active,
-        })
-
-    return templates.TemplateResponse("policies.html",
-        _ctx(request, session, active="policies", policies=policy_list))
-
-
-@router.get("/policies/create", response_class=HTMLResponse)
-async def policy_create_form(request: Request, db: AsyncSession = Depends(get_db)):
-    session = require_login(request)
-    if isinstance(session, RedirectResponse):
-        return session
-
-    orgs = (await db.execute(
-        select(OrganizationRecord).where(OrganizationRecord.status == "active").order_by(OrganizationRecord.org_id)
-    )).scalars().all()
-
-    return templates.TemplateResponse("policy_create.html",
-        _ctx(request, session, active="policies", form={}, orgs=orgs,
-             error=None, success=None))
-
-
-@router.post("/policies/create", response_class=HTMLResponse)
-async def policy_create_submit(request: Request, db: AsyncSession = Depends(get_db)):
-    session = require_login(request)
-    if isinstance(session, RedirectResponse):
-        return session
-    if not await verify_csrf(request, session):
-        return RedirectResponse(url="/dashboard/policies", status_code=303)
-
-    form_data = await request.form()
-    org_id       = form_data.get("org_id", "").strip()
-    target_org   = form_data.get("target_org_id", "").strip()
-    caps_raw     = form_data.get("capabilities", "").strip()
-    effect       = form_data.get("effect", "allow").strip()
-
-    form = {"org_id": org_id, "target_org_id": target_org, "capabilities": caps_raw, "effect": effect}
-
-    orgs = (await db.execute(
-        select(OrganizationRecord).where(OrganizationRecord.status == "active").order_by(OrganizationRecord.org_id)
-    )).scalars().all()
-
-    # Org user can only create policies for their own org
-    if not session.is_admin and org_id != session.org_id:
-        return templates.TemplateResponse("policy_create.html",
-            _ctx(request, session, active="policies", form=form, orgs=orgs,
-                 error="You can only create policies for your own organization.", success=None))
-
-    if not org_id or not target_org:
-        return templates.TemplateResponse("policy_create.html",
-            _ctx(request, session, active="policies", form=form, orgs=orgs,
-                 error="Organization and target organization are required.", success=None))
-
-    if org_id == target_org:
-        return templates.TemplateResponse("policy_create.html",
-            _ctx(request, session, active="policies", form=form, orgs=orgs,
-                 error="Organization and target must be different.", success=None))
-
-    caps = [c.strip() for c in caps_raw.split(",") if c.strip()] if caps_raw else []
-
-    policy_id = f"{org_id}::session-{target_org}-v1"
-    conditions: dict = {"target_org_id": [target_org]}
-    if caps:
-        conditions["capabilities"] = caps
-
-    existing = await get_policy(db, policy_id)
-    if existing:
-        return templates.TemplateResponse("policy_create.html",
-            _ctx(request, session, active="policies", form=form, orgs=orgs,
-                 error=f"Policy '{policy_id}' already exists.", success=None))
-
-    await create_policy(db, policy_id, org_id, "session", {"effect": effect, "conditions": conditions})
-    await log_event(db, "policy.created", "ok", org_id=org_id,
-                    details={"policy_id": policy_id, "target_org": target_org, "source": "dashboard"})
-
-    return templates.TemplateResponse("policy_create.html",
-        _ctx(request, session, active="policies", form={}, orgs=orgs, error=None,
-             success=f"Policy '{policy_id}' created. {org_id} → {target_org} [{effect}]"))
-
-
-@router.post("/policies/{policy_id:path}/deactivate", response_class=HTMLResponse)
-async def policy_deactivate(request: Request, policy_id: str, db: AsyncSession = Depends(get_db)):
-    session = require_login(request)
-    if isinstance(session, RedirectResponse):
-        return session
-    if not await verify_csrf(request, session):
-        return RedirectResponse(url="/dashboard/policies", status_code=303)
-
-    record = await get_policy(db, policy_id)
-    if record:
-        if not session.is_admin and record.org_id != session.org_id:
-            return RedirectResponse(url="/dashboard/policies", status_code=303)
-        await deactivate_policy(db, policy_id)
-        await log_event(db, "policy.deactivated", "ok", org_id=record.org_id,
-                        details={"policy_id": policy_id, "source": "dashboard"})
-
-    return RedirectResponse(url="/dashboard/policies", status_code=303)
-
-
-# ─────────────────────────────────────────────────────────────────────────────
-# Bindings — registry of approved org ↔ agent grants (cross-org)
+# Policy routes (/dashboard/policies + create + deactivate) moved to
+# ``app/dashboard/policies_routes.py`` since F-B-202 PR-8.
 #
-# A binding is the per-org authorization to let an external agent call into
-# this org. Three knobs the network admin needs at a glance:
-#   • status: pending / approved / revoked (lifecycle)
-#   • scope:  subset of the agent's declared capabilities that this org
-#             actually concedes — can be tighter than what the agent offers
-#   • org_id ↔ agent_id pair: bindings are directional, so A↔B requires two
-#             separate rows; revoking one keeps the other live, which is
-#             how an operator pins one-way traffic
-# ─────────────────────────────────────────────────────────────────────────────
-
-@router.get("/bindings", response_class=HTMLResponse)
-async def bindings_list(
-    request: Request,
-    db: AsyncSession = Depends(get_db),
-    status_filter: str | None = None,
-):
-    session = require_login(request)
-    if isinstance(session, RedirectResponse):
-        return session
-
-    bindings = await list_all_bindings(db, status_filter=status_filter)
-
-    # Hydrate each row with the underlying agent's capabilities so the
-    # scope editor can render a multi-select with the legal options.
-    from app.registry.store import get_agent_by_id
-    enriched = []
-    for b in bindings:
-        agent = await get_agent_by_id(db, b.agent_id)
-        enriched.append({
-            "id": b.id,
-            "org_id": b.org_id,
-            "agent_id": b.agent_id,
-            "status": b.status,
-            "scope": b.scope,
-            "approved_by": b.approved_by,
-            "approved_at": b.approved_at,
-            "created_at": b.created_at,
-            "agent_capabilities": list(agent.capabilities) if agent else [],
-            "agent_org_id": agent.org_id if agent else None,
-        })
-
-    counts = {
-        "pending":  sum(1 for b in bindings if b.status == "pending"),
-        "approved": sum(1 for b in bindings if b.status == "approved"),
-        "revoked":  sum(1 for b in bindings if b.status == "revoked"),
-    }
-
-    return templates.TemplateResponse(
-        "bindings.html",
-        _ctx(
-            request, session, active="bindings",
-            bindings=enriched,
-            counts=counts,
-            current_filter=status_filter or "",
-        ),
-    )
-
-
-@router.post("/bindings/{binding_id}/approve", response_class=HTMLResponse)
-async def bindings_approve(
-    request: Request, binding_id: int, db: AsyncSession = Depends(get_db),
-):
-    session = require_login(request)
-    if isinstance(session, RedirectResponse):
-        return session
-    if not await verify_csrf(request, session):
-        raise HTTPException(status_code=403, detail="Invalid CSRF token")
-
-    binding = await get_binding(db, binding_id)
-    if not binding:
-        raise HTTPException(status_code=404, detail="Binding not found")
-
-    approved = await approve_binding(
-        db, binding_id, approved_by="dashboard-network-admin",
-    )
-    if approved is None:
-        return RedirectResponse(url="/dashboard/bindings", status_code=303)
-
-    await log_event(
-        db, "binding.approved", "ok",
-        agent_id=binding.agent_id, org_id=binding.org_id,
-        details={"binding_id": binding_id, "source": "dashboard"},
-    )
-    return RedirectResponse(url="/dashboard/bindings", status_code=303)
-
-
-@router.post("/bindings/{binding_id}/revoke", response_class=HTMLResponse)
-async def bindings_revoke(
-    request: Request, binding_id: int, db: AsyncSession = Depends(get_db),
-):
-    session = require_login(request)
-    if isinstance(session, RedirectResponse):
-        return session
-    if not await verify_csrf(request, session):
-        raise HTTPException(status_code=403, detail="Invalid CSRF token")
-
-    binding = await get_binding(db, binding_id)
-    if not binding:
-        raise HTTPException(status_code=404, detail="Binding not found")
-
-    revoked = await revoke_binding(db, binding_id)
-    if revoked is None:
-        return RedirectResponse(url="/dashboard/bindings", status_code=303)
-
-    # Mirror the API path side effects: close active sessions + drop ws.
-    from app.broker.session import get_session_store
-    from app.broker.persistence import save_session
-    from app.broker.ws_manager import ws_manager
-
-    store = get_session_store()
-    closed = store.close_all_for_agent(revoked.agent_id)
-    for s in closed:
-        await save_session(db, s)
-    if ws_manager.is_connected(revoked.agent_id):
-        await ws_manager.disconnect(revoked.agent_id)
-
-    await log_event(
-        db, "binding.revoked", "ok",
-        agent_id=revoked.agent_id, org_id=revoked.org_id,
-        details={
-            "binding_id": binding_id, "source": "dashboard",
-            "sessions_closed": len(closed),
-        },
-    )
-    return RedirectResponse(url="/dashboard/bindings", status_code=303)
-
-
-@router.post("/bindings/{binding_id}/scope", response_class=HTMLResponse)
-async def bindings_update_scope(
-    request: Request, binding_id: int, db: AsyncSession = Depends(get_db),
-):
-    session = require_login(request)
-    if isinstance(session, RedirectResponse):
-        return session
-    if not await verify_csrf(request, session):
-        raise HTTPException(status_code=403, detail="Invalid CSRF token")
-
-    binding = await get_binding(db, binding_id)
-    if not binding:
-        raise HTTPException(status_code=404, detail="Binding not found")
-
-    form = await request.form()
-    raw_scope = form.getlist("scope")
-    new_scope = [s.strip() for s in raw_scope if s.strip()]
-
-    # Subset check against the agent's declared capabilities — same gate
-    # the API enforces on POST /v1/registry/bindings.
-    from app.registry.store import get_agent_by_id
-    agent = await get_agent_by_id(db, binding.agent_id)
-    if agent is None:
-        raise HTTPException(
-            status_code=404,
-            detail=f"Agent {binding.agent_id} not found — cannot validate scope",
-        )
-    agent_caps = set(agent.capabilities)
-    invalid = [s for s in new_scope if s not in agent_caps]
-    if invalid:
-        raise HTTPException(
-            status_code=400,
-            detail=(
-                f"Scope {invalid} not in agent capabilities: "
-                f"{sorted(agent_caps)}"
-            ),
-        )
-
-    await update_binding_scope(db, binding_id, new_scope)
-    await log_event(
-        db, "binding.scope_updated", "ok",
-        agent_id=binding.agent_id, org_id=binding.org_id,
-        details={
-            "binding_id": binding_id,
-            "old_scope": binding.scope,
-            "new_scope": new_scope,
-            "source": "dashboard",
-        },
-    )
-    return RedirectResponse(url="/dashboard/bindings", status_code=303)
+# Binding routes (/dashboard/bindings + approve / revoke / scope) moved
+# to ``app/dashboard/bindings_routes.py`` since F-B-202 PR-8.
